Moved to cassandra as it is the only caller
