+++ zipkin-storage/cassandra-v1/src/main/java/zipkin2/storage/cassandra/v1/SessionFactory.java
@@ -24,12 +24,12 @@ import com.datastax.driver.core.policies.RoundRobinPolicy;
 import com.datastax.driver.core.policies.TokenAwarePolicy;
 import com.google.common.collect.Sets;
 import com.google.common.io.Closer;
-import zipkin2.internal.HostAndPort;
 import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
+import zipkin2.storage.cassandra.internal.HostAndPort;
 
 /**
  * Creates a session and ensures schema if configured. Closes the cluster and session if any
@@ -100,8 +100,8 @@ public interface SessionFactory {
     static List<InetSocketAddress> parseContactPoints(CassandraStorage cassandra) {
       List<InetSocketAddress> result = new ArrayList<>();
       for (String contactPoint : cassandra.contactPoints.split(",")) {
-        HostAndPort parsed = HostAndPort.fromString(contactPoint);
-        result.add(new InetSocketAddress(parsed.getHost(), parsed.getPortOrDefault(9042)));
+        HostAndPort parsed = HostAndPort.fromString(contactPoint, 9042);
+        result.add(new InetSocketAddress(parsed.getHost(), parsed.getPort()));
       }
       return result;
     }
